Add Opting out data

+++ uninstall.php
@@ -2,6 +2,7 @@
 
 /**
  * Fired when the plugin is uninstalled.
+ *
  * @link       https://www.expresstechsoftwares.com
  * @since      1.0.0
  *
@@ -17,8 +18,11 @@ if ( $_REQUEST['plugin'] == 'connect-learnpress-discord-addon/learnpress-discord
 		&& $_REQUEST['slug'] == 'learnpress-discord-add'
 	&& wp_verify_nonce( $_REQUEST['_ajax_nonce'], 'updates' )
   ) {
-	global $wpdb;
-	  $wpdb->query( 'DELETE FROM ' . $wpdb->prefix . "usermeta WHERE `meta_key` LIKE '_ets_learnpress_discord%'" );
-	  $wpdb->query( 'DELETE FROM ' . $wpdb->prefix . "options WHERE `option_name` LIKE 'ets_learnpress_discord_%'" );
+	$ets_learnpress_discord_data_erases = sanitize_text_field( trim( get_option( 'ets_learnpress_discord_data_erases' ) ) );
+	if ( $ets_learnpress_discord_data_erases ) {
+		global $wpdb;
+		$wpdb->query( 'DELETE FROM ' . $wpdb->prefix . "usermeta WHERE `meta_key` LIKE '_ets_learnpress_discord%'" );
+		$wpdb->query( 'DELETE FROM ' . $wpdb->prefix . "options WHERE `option_name` LIKE 'ets_learnpress_discord_%'" );
+	}
 }
